+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Blue Triangle 3.15.11, Apr 30 2026
+### Bug Fixes and Improvements
+- Fixed an occasional concurrency crash while enabling signal crash feature.
+
 # Blue Triangle 3.15.10, Mar 23 2026
 ### Bug Fixes and Improvements
 - Added breadcrumbs for ANRs, crashes, and memory warnings